Config window done for now
Starting Load Save - App Should monopolize the event trigger load / save.

Thinking: FS should send an event to initialize their JSON_Object to the root node. This event will send the root node, but with that we don't need access to root node and we can make thigns multithreaded later maybe. Just that is a bottleneck.